+++ python/src/scippneutron/conversion.py
